requirements.txt
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
requests==2.31.0
jsonschema==4.20.0
faker==20.1.0
pydantic==2.5.0
python-dotenv==1.0.0
allure-pytest==2.13.2
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-json-report==1.5.0 
